+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -2,14 +2,11 @@ package ru.yandex.practicum.filmorate.controllers;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.web.bind.annotation.*;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 
 import javax.validation.Valid;
-import java.time.LocalDate;
 import java.util.List;
 
 @RestController
@@ -21,13 +18,13 @@ public class FilmController {
 
     @PostMapping
     private Film addFilm(@Valid @RequestBody Film film) {
-        validate(film);
+        log.info("Film '" + film.getName() + "' added successfully");
         return service.addFilm(film);
     }
 
     @PutMapping
     public Film updateFilm(@Valid @RequestBody Film film) {
-        validate(film);
+        log.info("Movie data update: {}", film);
         return service.updateFilm(film);
     }
 
@@ -56,27 +53,5 @@ public class FilmController {
     @GetMapping("/popular")
     public List<Film> getMostLikedFilms(@RequestParam(required = false, defaultValue = "10") Integer count) {
         return service.getMostLikedFilms(count);
-    }
-
-    public static void validate(Film film) {
-        if (StringUtils.isBlank(film.getName())) {
-            log.error("Validation error: Films name can't be blank");
-            throw new ValidationException("Films name can't be blank");
-        }
-
-        if (film.getDescription().length() > 200) {
-            log.error("Validation error: Films description can't be longer than 200 chars");
-            throw new ValidationException("Films description can't be longer than 200 chars");
-        }
-
-        if (film.getReleaseDate().isBefore(LocalDate.of(1895, 12, 28))) {
-            log.error("Validation error: Films release date can't be before 28.12.1895");
-            throw new ValidationException("Films release date can't be before 28.12.1895");
-        }
-
-        if (film.getDuration() < 0) {
-            log.error("Validation error: Films duration can't be negative");
-            throw new ValidationException("Films duration can't be negative");
-        }
     }
 }
